+++ .prettierrc
@@ -1,0 +1,4 @@
+{
+  "trailingComma": "es5",
+  "arrowParens": "always"
+}
