+++ package.json
@@ -5,8 +5,8 @@
   "main": "index.js",
   "scripts": {
     "start": "node ./src/index.js",
-    "dev": "nodedemon ./src/index.js",
-    "test": "mocha spec && yarn lint",
+    "dev": "nodemon ./src/index.js",
+    "test": "mocha test && yarn lint",
     "lint": "eslint src --fix && prettier --write src"
   },
   "keywords": [],
